[#24] Update partContentLength in S3_upload_part to int64_t.

+++ inc/libs3.h
@@ -2552,7 +2552,7 @@ void S3_initiate_multipart(S3BucketContext *bucketContext, const char *key,
 void S3_upload_part(S3BucketContext *bucketContext, const char *key,
                     S3PutProperties * putProperties,
                     S3PutObjectHandler *handler,
-                    int seq, const char *upload_id, int partContentLength,
+                    int seq, const char *upload_id, int64_t partContentLength,
                     S3RequestContext *requestContext,
                     int timeoutMs,
                     void *callbackData);
